доделала модуляцию
